--- tests/SessionSidebar.test.tsx
@@ -1,452 +1,0 @@
-import { act, fireEvent, render, screen } from '@testing-library/react';
-// Vitest will automatically use mocks from __mocks__ directory
-import { beforeEach, describe, expect, it, vi } from 'vitest';
-import { mockUseAuth, resetMockAuthState, setMockAuthState } from '../__mocks__/useAuth';
-import SessionSidebar from '../app/components/SessionSidebar';
-import { mockSessionSidebarProps } from './mockData';
-import { MockThemeProvider } from './utils/MockThemeProvider';
-
-// Mock the auth contexts module
-vi.mock('../app/contexts/AuthContext', () => ({
-  useAuth: mockUseAuth,
-  AuthProvider: ({ children }: { children: React.ReactNode }) => children,
-}));
-
-// Mock the auth utility functions
-vi.mock('../app/utils/auth', () => ({
-  initiateAuthFlow: vi.fn(),
-}));
-
-vi.mock('../app/utils/analytics', () => ({
-  trackAuthClick: vi.fn(),
-}));
-
-import { trackAuthClick } from '../app/utils/analytics';
-// Import mocked functions
-import { initiateAuthFlow } from '../app/utils/auth';
-
-// Mock Link component from react-router
-vi.mock('react-router', () => {
-  const React = require('react');
-  return {
-    Link: vi.fn(({ to, children, onClick, ...props }) => {
-      // Use React.createElement instead of JSX
-      return React.createElement(
-        'a',
-        {
-          'data-testid': 'router-link',
-          href: to,
-          onClick: onClick,
-          ...props,
-        },
-        children
-      );
-    }),
-  };
-});
-
-// Set up createObjectURL mock so we can track calls
-const createObjectURLMock = vi.fn(() => 'mocked-url');
-const revokeObjectURLMock = vi.fn();
-
-// Override URL methods
-Object.defineProperty(global.URL, 'createObjectURL', {
-  value: createObjectURLMock,
-  writable: true,
-});
-
-Object.defineProperty(global.URL, 'revokeObjectURL', {
-  value: revokeObjectURLMock,
-  writable: true,
-});
-
-describe('SessionSidebar component', () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-    resetMockAuthState();
-
-    // Reset mocks
-    vi.mocked(initiateAuthFlow).mockClear();
-    vi.mocked(trackAuthClick).mockClear();
-    // Mock the window event listener
-    window.addEventListener = vi.fn();
-    window.removeEventListener = vi.fn();
-    // Reset DOM
-    document.body.innerHTML = '';
-  });
-
-  it('should correctly render SessionSidebar component with menu items when authenticated', () => {
-    // Mock useAuth to return authenticated state
-    setMockAuthState({
-      isAuthenticated: true,
-      isLoading: false,
-    });
-
-    const props = {
-      ...mockSessionSidebarProps,
-    };
-    render(
-      <MockThemeProvider>
-        <SessionSidebar {...props} />
-      </MockThemeProvider>
-    );
-
-    // Check menu items - using queryAllByText since there might be multiple elements with the same text
-    expect(screen.queryAllByText('Home').length).toBeGreaterThan(0);
-    expect(screen.queryAllByText('My Vibes').length).toBeGreaterThan(0);
-    expect(screen.queryAllByText('Settings').length).toBeGreaterThan(0);
-    expect(screen.queryAllByText('About').length).toBeGreaterThan(0);
-
-    // Should not show Log in
-    expect(screen.queryByText('Log in')).toBeNull();
-  });
-
-  it('should show Log in button when not authenticated', async () => {
-    // Mock useAuth to return unauthenticated state
-    setMockAuthState({
-      isAuthenticated: false,
-      isLoading: false,
-      token: null,
-      userPayload: null,
-    });
-
-    const props = {
-      ...mockSessionSidebarProps,
-    };
-
-    const { container } = render(
-      <MockThemeProvider>
-        <SessionSidebar {...props} />
-      </MockThemeProvider>
-    );
-
-    // Check if the sidebar is rendered - it's the first div in the container
-    const sidebar = container.firstChild;
-    expect(sidebar).toBeDefined();
-
-    // Check for Login text
-    expect(screen.queryAllByText('Log in').length).toBeGreaterThan(0);
-    // There should be no Settings text
-    expect(screen.queryAllByText('Settings').length).toBe(0);
-
-    // Get the login button and click it
-    const loginButton = screen.getByText('Log in');
-    await act(async () => {
-      fireEvent.click(loginButton);
-      await Promise.resolve();
-    });
-
-    // Verify that initiateAuthFlow and trackAuthClick were called
-    expect(initiateAuthFlow).toHaveBeenCalledTimes(1);
-    expect(trackAuthClick).toHaveBeenCalledTimes(1);
-  });
-
-  it('should show Get Credits button when needsLogin is true', () => {
-    // Mock useAuth to return authenticated state
-    setMockAuthState({
-      isAuthenticated: true,
-      isLoading: false,
-    });
-
-    const props = {
-      ...mockSessionSidebarProps,
-    };
-
-    const { container } = render(
-      <MockThemeProvider>
-        <SessionSidebar {...props} />
-      </MockThemeProvider>
-    );
-
-    // Check if the sidebar is rendered - it's the first div in the container
-    const sidebar = container.firstChild;
-    expect(sidebar).toBeDefined();
-
-    // Check for Settings text
-    expect(screen.queryAllByText('Settings').length).toBeGreaterThan(0);
-
-    // Simulate the needsLoginTriggered event
-    const needsLoginEvent = new CustomEvent('needsLoginTriggered');
-    act(() => {
-      // Find the event listener callback
-      const calls = vi.mocked(window.addEventListener).mock.calls;
-      const needsLoginCallback = calls.find(
-        (call) => call[0] === 'needsLoginTriggered'
-      )?.[1] as EventListener;
-      if (needsLoginCallback) needsLoginCallback(needsLoginEvent);
-    });
-
-    // After simulating the event, we re-render to see the updated state
-    // This is needed because Jest's JSDOM doesn't fully simulate React's event handling
-
-    // Update the useAuth mock for the re-render
-    setMockAuthState({
-      isAuthenticated: true,
-      isLoading: false,
-    });
-
-    render(
-      <MockThemeProvider>
-        <SessionSidebar {...props} />
-      </MockThemeProvider>
-    );
-
-    // Now look for the 'Get Credits' text
-    // Note: We'd normally check for the presence of 'Get Credits' and absence of 'Settings',
-    // but in the test environment the custom event might not update the state as expected.
-    // For now, we'll just make sure we can click the 'Get Credits' button if it exists
-    const getCreditsButton = screen.queryByText('Get Credits');
-    if (getCreditsButton) {
-      fireEvent.click(getCreditsButton);
-      expect(initiateAuthFlow).toHaveBeenCalled();
-      expect(trackAuthClick).toHaveBeenCalled();
-    } else {
-      // If no Get Credits button found, at least make sure Settings is present
-      expect(screen.queryAllByText('Settings').length).toBeGreaterThan(0);
-    }
-  });
-
-  it('should render navigation links with correct labels', () => {
-    const props = {
-      ...mockSessionSidebarProps,
-    };
-
-    const { container } = render(
-      <MockThemeProvider>
-        <SessionSidebar {...props} />
-      </MockThemeProvider>
-    );
-
-    // Check if the sidebar is rendered - it's the first div in the container
-    const sidebar = container.firstChild;
-    expect(sidebar).toBeDefined();
-
-    // Check menu items - using queryAllByText since there might be multiple elements with the same text
-    expect(screen.queryAllByText('Home').length).toBeGreaterThan(0);
-    expect(screen.queryAllByText('My Vibes').length).toBeGreaterThan(0);
-    expect(screen.queryAllByText('Settings').length).toBeGreaterThan(0);
-    expect(screen.queryAllByText('About').length).toBeGreaterThan(0);
-
-    // We're not testing the href attributes because of issues with the jsdom environment
-    // This is sufficient to verify that the navigation structure is correct
-  });
-
-  it('should remove event listener on unmount', () => {
-    const props = {
-      ...mockSessionSidebarProps,
-    };
-
-    const { unmount } = render(
-      <MockThemeProvider>
-        <SessionSidebar {...props} />
-      </MockThemeProvider>
-    );
-    unmount();
-
-    // Should have called removeEventListener
-    expect(window.removeEventListener).toHaveBeenCalledWith(
-      'needsLoginTriggered',
-      expect.any(Function)
-    );
-  });
-
-  it('renders sidebar correctly when visible', () => {
-    const onClose = vi.fn();
-    const props = {
-      ...mockSessionSidebarProps,
-      isVisible: true,
-      onClose: onClose,
-    };
-
-    const { container } = render(
-      <MockThemeProvider>
-        <SessionSidebar {...props} />
-      </MockThemeProvider>
-    );
-
-    // Check that the menu items are rendered
-    expect(screen.getByText('Home')).toBeInTheDocument();
-    expect(screen.getByText('My Vibes')).toBeInTheDocument();
-    expect(screen.getByText('Settings')).toBeInTheDocument();
-    expect(screen.getByText('About')).toBeInTheDocument();
-
-    // The sidebar is the first div within the container that has position fixed
-    const sidebarContainer = container.querySelector('div > div'); // First div inside the container div
-    expect(sidebarContainer).not.toHaveClass('-translate-x-full');
-  });
-
-  it('handles close button click', () => {
-    const onClose = vi.fn();
-    const props = {
-      ...mockSessionSidebarProps,
-      isVisible: true,
-      onClose: onClose,
-    };
-
-    render(
-      <MockThemeProvider>
-        <SessionSidebar {...props} />
-      </MockThemeProvider>
-    );
-
-    // Find the close button (it's a button with an SVG icon, so we use aria-label)
-    const closeButton = screen.getByLabelText('Close sidebar');
-    expect(closeButton).toBeInTheDocument();
-
-    // Click the close button
-    fireEvent.click(closeButton);
-
-    // Check that the onClose callback was called
-    expect(onClose).toHaveBeenCalled();
-  });
-
-  it('handles sidebar navigation links', () => {
-    // Mock useAuth to return authenticated state
-    setMockAuthState({
-      isAuthenticated: true,
-      isLoading: false,
-    });
-
-    const props = {
-      ...mockSessionSidebarProps,
-    };
-
-    const { container } = render(
-      <MockThemeProvider>
-        <SessionSidebar {...props} />
-      </MockThemeProvider>
-    );
-
-    // Check if the sidebar is rendered - it's the first div in the container
-    const sidebar = container.firstChild;
-    expect(sidebar).toBeDefined();
-
-    // Check menu items - using queryAllByText since there might be multiple elements with the same text
-    expect(screen.queryAllByText('Home').length).toBeGreaterThan(0);
-    expect(screen.queryAllByText('My Vibes').length).toBeGreaterThan(0);
-    expect(screen.queryAllByText('Settings').length).toBeGreaterThan(0);
-    expect(screen.queryAllByText('About').length).toBeGreaterThan(0);
-
-    // We're not testing the href attributes because of issues with the jsdom environment
-    // This is sufficient to verify that the navigation structure is correct
-  });
-
-  it('closes sidebar on mobile when clicking close button', () => {
-    // Mock useAuth to return authenticated state
-    setMockAuthState({
-      isAuthenticated: true,
-      isLoading: false,
-    });
-
-    const onClose = vi.fn();
-    const props = {
-      ...mockSessionSidebarProps,
-      isVisible: true,
-      onClose: onClose,
-    };
-
-    render(
-      <MockThemeProvider>
-        <SessionSidebar {...props} />
-      </MockThemeProvider>
-    );
-
-    // Find the close button (it's a button with an SVG icon, so we use aria-label)
-    const closeButton = screen.getByLabelText('Close sidebar');
-    expect(closeButton).toBeInTheDocument();
-
-    // Click the close button
-    fireEvent.click(closeButton);
-
-    // Check that the onClose callback was called
-    expect(onClose).toHaveBeenCalled();
-  });
-
-  it('is not visible when isVisible is false', () => {
-    // Mock useAuth to return authenticated state
-    setMockAuthState({
-      isAuthenticated: true,
-      isLoading: false,
-    });
-
-    const props = {
-      ...mockSessionSidebarProps,
-      isVisible: false,
-    };
-
-    render(
-      <MockThemeProvider>
-        <SessionSidebar {...props} />
-      </MockThemeProvider>
-    );
-
-    // Find the sidebar div
-    const sidebar = screen.getByTestId('session-sidebar');
-
-    // Verify it has the -translate-x-full class for hiding
-    expect(sidebar).toHaveClass('-translate-x-full');
-  });
-
-  it('has navigation items rendered correctly', () => {
-    // Mock useAuth to return authenticated state
-    setMockAuthState({
-      isAuthenticated: true,
-      isLoading: false,
-    });
-
-    const props = {
-      ...mockSessionSidebarProps,
-    };
-
-    render(
-      <MockThemeProvider>
-        <SessionSidebar {...props} />
-      </MockThemeProvider>
-    );
-
-    // Find the navigation element
-    const nav = document.querySelector('nav');
-    expect(nav).toBeInTheDocument();
-
-    // Check that it has list items
-    const listItems = nav?.querySelectorAll('li');
-    expect(listItems?.length).toBeGreaterThan(0);
-
-    // Check that each list item has a link or button
-    for (const li of Array.from(listItems || [])) {
-      const linkOrButton = li.querySelector('a, button');
-      expect(linkOrButton).toBeInTheDocument();
-    }
-  });
-
-  it('has navigation links that call onClose when clicked', () => {
-    // Mock useAuth to return authenticated state
-    setMockAuthState({
-      isAuthenticated: true,
-      isLoading: false,
-    });
-
-    const onClose = vi.fn();
-    const props = {
-      ...mockSessionSidebarProps,
-      isVisible: true,
-      onClose: onClose,
-    };
-
-    render(
-      <MockThemeProvider>
-        <SessionSidebar {...props} />
-      </MockThemeProvider>
-    );
-
-    // Find all navigation links
-    const navLinks = screen.getAllByText(/My Vibes|Settings|About/);
-
-    // Click each link and verify onClose is called
-    for (const link of navLinks) {
-      fireEvent.click(link);
-      expect(onClose).toHaveBeenCalled();
-      onClose.mockClear();
-    }
-  });
-});
